Move Run Inspector into `inspector/run` dir.
